Set up classes
Issue #281 Set up classes for generic syllable recognition.
