+++ src/AudioAnalysisTools/AudioAnalysisTools.csproj
@@ -310,6 +310,7 @@
     <Compile Include="LongDurationSpectrograms\Zooming\ZoomFocusedSpectrograms.cs" />
     <Compile Include="Oscillations2014.cs" />
     <Compile Include="Oscillations2019.cs" />
+    <Compile Include="RecognizerTools.cs" />
     <Compile Include="RidgeDetection.cs" />
     <Compile Include="SiteDescription.cs" />
     <Compile Include="StandardSpectrograms\AmplitudeSpectrogram.cs" />
@@ -426,6 +427,7 @@
     <Analyzer Include="..\..\packages\StyleCop.Analyzers.Unstable.1.1.1.61\analyzers\dotnet\cs\StyleCop.Analyzers.CodeFixes.dll" />
     <Analyzer Include="..\..\packages\StyleCop.Analyzers.Unstable.1.1.1.61\analyzers\dotnet\cs\StyleCop.Analyzers.dll" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
